Created todo_list_assignement_with_each file
Creating a class identical to the todo_list_assignment.rb assignment. The TodoList class has the TodoList#each method invoked with the { |todo| puts todo } block made implicitly available. The each method fetches the element at the position of the passed in argument, and yields to a block, passing in that element as a block parameter.