Update todo_list_assignment_with_select.rb
Changed the TodoList#select to imitate the Array#select method, where the Array#select method returns the calling object. Previously, the TodoList#select method returned an Array object, which is not the same type as TodoList. The code changes in this commit make this TodoList#select method return a TodoList object, rather than an array object. This is closer to the behavior of the Array#select method.