Create final_todolist_assignment.rb
Implemented the following methods:
  1. find_by_title(search_title)
  2. all_done()
  3. all_not_done()
  4. mark_done()
  5. mark_all_done()
  6. mark_all_undone()

Made Changes to find_by_title():
The requirements specify to return the FIRST element that matches the search title. Otherwise, the method needs to return false. Previously, just the TodoList#select method was called, which would return an array (Either an empty array if nothing was found, or an array containing the found elements). The output for the TodoList#select method does not exactly meet the requirements, so the Array#first method is called, which returns the first element of an array, or returns nil if the array is empty. This method chaining functionality for TodoList#select#first satisfies the requirements, so this implementation was included in the final draft.